Moving syncer code to another file
